pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amadeus>=12.0.0",
    "langchain>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.32",
    "openai>=1.104.2",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "tabulate>=0.9.0",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "redis>=5.2.2",
]
